requirements.txt
accelerate==0.32.1
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
diffusers==0.29.2
filelock==3.15.4
fsspec==2024.6.1
google-api-core==2.20.0
google-auth==2.32.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
huggingface-hub==0.24.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mcp==1.2.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
Pillow==10.4.0
prompt-toolkit==3.0.47
protobuf==5.27.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
sympy==1.13.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13